Moved controllers to their own folder and updated require statements
