+++ server/routes/frontend.js
@@ -3,11 +3,11 @@
 // serve up this file at any '/' route, which should go through this list and see if anything applies
 // then, it should proceed to the next routes.
 
-var express = require('express');
-var router = express.Router();
+const express = require('express');
+const router = express.Router();
 const { Comment, User } = require ('../../db/pg-index.js');
-const { create } = require('../controller.js');
-const { getDescription, getUser, getUserId, getCommentsForVideo, getAllDescriptions } = require('../frontend-controller.js');
+const { create } = require('../controllers/crud.js');
+const { getDescription, getUser, getUserId, getCommentsForVideo, getAllDescriptions } = require('../controllers/frontend.js');
 
 
 /* ROUTE 1
